Moving handlers to better location.
