+++ pkg/tasks/handlers.go
@@ -1,4 +1,4 @@
-package main
+package tasks
 
 import (
 	"context"
